+++ Source/LupusCore/CryptoHash.h
@@ -24,30 +24,30 @@ namespace Lupus {
                     return std::make_shared<CryptoHash<T>>();
                 }
 
-                virtual uint32_t BlockSize() const NOEXCEPT override
+                virtual size_t BlockSize() const NOEXCEPT override
                 {
                     return T::BLOCKSIZE * 8;
                 }
 
-                virtual uint32_t HashSize() const NOEXCEPT override
+                virtual size_t HashSize() const NOEXCEPT override
                 {
                     return T::DIGESTSIZE * 8;
                 }
 
                 virtual std::vector<uint8_t> ComputeHash(const std::vector<uint8_t>& buffer) throw(std::out_of_range) override
                 {
-                    return ComputeHash(buffer, 0, (uint32_t)buffer.size());
+                    return ComputeHash(buffer, 0, (size_t)buffer.size());
                 }
 
                 virtual std::vector<uint8_t> ComputeHash(std::shared_ptr<Stream> stream) throw(std::out_of_range) override
                 {
-                    uint32_t size = (uint32_t)(stream->Length() - stream->Position());
+                    size_t size = (size_t)(stream->Length() - stream->Position());
                     std::vector<uint8_t> buffer(size);
                     stream->Read(buffer, 0, size);
                     return ComputeHash(buffer, 0, size);
                 }
 
-                virtual std::vector<uint8_t> ComputeHash(const std::vector<uint8_t>& input, uint32_t inputOffset, uint32_t inputCount) throw(std::out_of_range) override
+                virtual std::vector<uint8_t> ComputeHash(const std::vector<uint8_t>& input, size_t inputOffset, size_t inputCount) throw(std::out_of_range) override
                 {
                     static const int digestsize = T::DIGESTSIZE;
 
@@ -79,30 +79,30 @@ namespace Lupus {
                     return std::make_shared<CryptoHash<CryptoPP::SHA3_224>>();
                 }
 
-                virtual uint32_t BlockSize() const NOEXCEPT override
+                virtual size_t BlockSize() const NOEXCEPT override
                 {
                     return mAlgorithm.BlockSize();
                 }
 
-                virtual uint32_t HashSize() const NOEXCEPT override
+                virtual size_t HashSize() const NOEXCEPT override
                 {
                     return mAlgorithm.DigestSize();
                 }
 
                 virtual std::vector<uint8_t> ComputeHash(const std::vector<uint8_t>& buffer) throw(std::out_of_range) override
                 {
-                    return ComputeHash(buffer, 0, (uint32_t)buffer.size());
+                    return ComputeHash(buffer, 0, (size_t)buffer.size());
                 }
 
                 virtual std::vector<uint8_t> ComputeHash(std::shared_ptr<Stream> stream) throw(std::out_of_range) override
                 {
-                    uint32_t size = (uint32_t)(stream->Length() - stream->Position());
+                    size_t size = (size_t)(stream->Length() - stream->Position());
                     std::vector<uint8_t> buffer(size);
                     stream->Read(buffer, 0, size);
                     return ComputeHash(buffer, 0, size);
                 }
 
-                virtual std::vector<uint8_t> ComputeHash(const std::vector<uint8_t>& input, uint32_t inputOffset, uint32_t inputCount) throw(std::out_of_range) override
+                virtual std::vector<uint8_t> ComputeHash(const std::vector<uint8_t>& input, size_t inputOffset, size_t inputCount) throw(std::out_of_range) override
                 {
                     if (inputOffset > input.size() || inputCount > input.size() - inputOffset) {
                         throw std::out_of_range("offset and size does not match buffer size");
@@ -132,30 +132,30 @@ namespace Lupus {
                     return std::make_shared<CryptoHash<CryptoPP::SHA3_256>>();
                 }
 
-                virtual uint32_t BlockSize() const NOEXCEPT override
+                virtual size_t BlockSize() const NOEXCEPT override
                 {
                     return mAlgorithm.BlockSize();
                 }
 
-                virtual uint32_t HashSize() const NOEXCEPT override
+                virtual size_t HashSize() const NOEXCEPT override
                 {
                     return mAlgorithm.DigestSize();
                 }
 
                 virtual std::vector<uint8_t> ComputeHash(const std::vector<uint8_t>& buffer) throw(std::out_of_range) override
                 {
-                    return ComputeHash(buffer, 0, (uint32_t)buffer.size());
+                    return ComputeHash(buffer, 0, (size_t)buffer.size());
                 }
 
                 virtual std::vector<uint8_t> ComputeHash(std::shared_ptr<Stream> stream) throw(std::out_of_range) override
                 {
-                    uint32_t size = (uint32_t)(stream->Length() - stream->Position());
+                    size_t size = (size_t)(stream->Length() - stream->Position());
                     std::vector<uint8_t> buffer(size);
                     stream->Read(buffer, 0, size);
                     return ComputeHash(buffer, 0, size);
                 }
 
-                virtual std::vector<uint8_t> ComputeHash(const std::vector<uint8_t>& input, uint32_t inputOffset, uint32_t inputCount) throw(std::out_of_range) override
+                virtual std::vector<uint8_t> ComputeHash(const std::vector<uint8_t>& input, size_t inputOffset, size_t inputCount) throw(std::out_of_range) override
                 {
                     if (inputOffset > input.size() || inputCount > input.size() - inputOffset) {
                         throw std::out_of_range("offset and size does not match buffer size");
@@ -185,30 +185,30 @@ namespace Lupus {
                     return std::make_shared<CryptoHash<CryptoPP::SHA3_384>>();
                 }
 
-                virtual uint32_t BlockSize() const NOEXCEPT override
+                virtual size_t BlockSize() const NOEXCEPT override
                 {
                     return mAlgorithm.BlockSize();
                 }
 
-                virtual uint32_t HashSize() const NOEXCEPT override
+                virtual size_t HashSize() const NOEXCEPT override
                 {
                     return mAlgorithm.DigestSize();
                 }
 
                 virtual std::vector<uint8_t> ComputeHash(const std::vector<uint8_t>& buffer) throw(std::out_of_range) override
                 {
-                    return ComputeHash(buffer, 0, (uint32_t)buffer.size());
+                    return ComputeHash(buffer, 0, (size_t)buffer.size());
                 }
 
                 virtual std::vector<uint8_t> ComputeHash(std::shared_ptr<Stream> stream) throw(std::out_of_range) override
                 {
-                    uint32_t size = (uint32_t)(stream->Length() - stream->Position());
+                    size_t size = (size_t)(stream->Length() - stream->Position());
                     std::vector<uint8_t> buffer(size);
                     stream->Read(buffer, 0, size);
                     return ComputeHash(buffer, 0, size);
                 }
 
-                virtual std::vector<uint8_t> ComputeHash(const std::vector<uint8_t>& input, uint32_t inputOffset, uint32_t inputCount) throw(std::out_of_range) override
+                virtual std::vector<uint8_t> ComputeHash(const std::vector<uint8_t>& input, size_t inputOffset, size_t inputCount) throw(std::out_of_range) override
                 {
                     if (inputOffset > input.size() || inputCount > input.size() - inputOffset) {
                         throw std::out_of_range("offset and size does not match buffer size");
@@ -238,30 +238,30 @@ namespace Lupus {
                     return std::make_shared<CryptoHash<CryptoPP::SHA3_512>>();
                 }
 
-                virtual uint32_t BlockSize() const NOEXCEPT override
+                virtual size_t BlockSize() const NOEXCEPT override
                 {
                     return mAlgorithm.BlockSize();
                 }
 
-                virtual uint32_t HashSize() const NOEXCEPT override
+                virtual size_t HashSize() const NOEXCEPT override
                 {
                     return mAlgorithm.DigestSize();
                 }
 
                 virtual std::vector<uint8_t> ComputeHash(const std::vector<uint8_t>& buffer) throw(std::out_of_range) override
                 {
-                    return ComputeHash(buffer, 0, (uint32_t)buffer.size());
+                    return ComputeHash(buffer, 0, (size_t)buffer.size());
                 }
 
                 virtual std::vector<uint8_t> ComputeHash(std::shared_ptr<Stream> stream) throw(std::out_of_range) override
                 {
-                    uint32_t size = (uint32_t)(stream->Length() - stream->Position());
+                    size_t size = (size_t)(stream->Length() - stream->Position());
                     std::vector<uint8_t> buffer(size);
                     stream->Read(buffer, 0, size);
                     return ComputeHash(buffer, 0, size);
                 }
 
-                virtual std::vector<uint8_t> ComputeHash(const std::vector<uint8_t>& input, uint32_t inputOffset, uint32_t inputCount) throw(std::out_of_range) override
+                virtual std::vector<uint8_t> ComputeHash(const std::vector<uint8_t>& input, size_t inputOffset, size_t inputCount) throw(std::out_of_range) override
                 {
                     if (inputOffset > input.size() || inputCount > input.size() - inputOffset) {
                         throw std::out_of_range("offset and size does not match buffer size");
