+++ Source/LupusCore/IPAddress.h
@@ -30,7 +30,7 @@ namespace Lupus {
                  *
                  * \param[in]   ipv4    Ganzzahl die eine IPv4 Adresse beinhaltet.
                  */
-                explicit IPAddress(uint32_t ipv4) NOEXCEPT;
+                explicit IPAddress(size_t ipv4) NOEXCEPT;
 
                 /*!
                  * Dieser Konstruktor ruft IPAddress(address, 0) auf.
@@ -52,10 +52,10 @@ namespace Lupus {
                  * \param[in]   ipv6    Die IPv6 Adresse in Netzwerkformat.
                  * \param[in]   scopeid Der Scope Identifier der IPv6 Adresse.
                  */
-                IPAddress(const std::vector<uint8_t>& ipv6, uint32_t scopeid) throw(std::length_error);
+                IPAddress(const std::vector<uint8_t>& ipv6, size_t scopeid) throw(std::length_error);
 
                 /*!
-                 * \sa IPAddress::IPAddress(const std::vector<uint8_t>&, uint32_t)
+                 * \sa IPAddress::IPAddress(const std::vector<uint8_t>&, size_t)
                  */
                 IPAddress(std::initializer_list<uint8_t> ilist) throw(std::length_error);
                 virtual ~IPAddress() = default;
@@ -90,14 +90,14 @@ namespace Lupus {
                 /*!
                  * \returns Den Scope Identifier der IPv6 Adresse.
                  */
-                virtual uint32_t ScopeId() const throw(socket_error);
+                virtual size_t ScopeId() const throw(socket_error);
 
                 /*!
                  * Setzt den Scope Identifier der IPv6 Adresse.
                  *
                  * \param[in]   value   Der neue Wert des Scope Identifiers.
                  */
-                virtual void ScopeId(uint32_t value) throw(socket_error);
+                virtual void ScopeId(size_t value) throw(socket_error);
 
                 /*!
                  * \returns Das Präsentationsformat der IP-Adresse.
@@ -160,7 +160,7 @@ namespace Lupus {
 
                 AddressFamily mFamily;
                 std::vector<uint8_t> mAddress;
-                uint32_t mScopeId = 0;
+                size_t mScopeId = 0;
             };
         }
     }
