+++ Source/LupusCore/MemoryStream.cpp
@@ -11,7 +11,7 @@ namespace Lupus {
         mIterator = begin(mBuffer);
     }
 
-    MemoryStream::MemoryStream(uint32_t length)
+    MemoryStream::MemoryStream(size_t length)
     {
         mBuffer = vector<uint8_t>(length);
         mIterator = begin(mBuffer);
@@ -24,7 +24,7 @@ namespace Lupus {
         mWritable = canWrite;
     }
 
-    MemoryStream::MemoryStream(const vector<uint8_t>& buffer, uint32_t offset, uint32_t size)
+    MemoryStream::MemoryStream(const vector<uint8_t>& buffer, size_t offset, size_t size)
     {
         if (offset > buffer.size() || size > buffer.size() - offset) {
             throw out_of_range("offset and size does not match buffer size");
@@ -34,7 +34,7 @@ namespace Lupus {
         mIterator = begin(mBuffer);
     }
 
-    MemoryStream::MemoryStream(const vector<uint8_t>& buffer, uint32_t offset, uint32_t size, bool canWrite)
+    MemoryStream::MemoryStream(const vector<uint8_t>& buffer, size_t offset, size_t size, bool canWrite)
     {
         if (offset > buffer.size() || size > buffer.size() - offset) {
             throw out_of_range("offset and size does not match buffer size");
@@ -45,7 +45,7 @@ namespace Lupus {
         mIterator = begin(mBuffer);
     }
 
-    MemoryStream::MemoryStream(const vector<uint8_t>& buffer, uint32_t offset, uint32_t size, bool canWrite, bool visible)
+    MemoryStream::MemoryStream(const vector<uint8_t>& buffer, size_t offset, size_t size, bool canWrite, bool visible)
     {
         if (offset > buffer.size() || size > buffer.size() - offset) {
             throw out_of_range("offset and size does not match buffer size");
@@ -88,7 +88,7 @@ namespace Lupus {
             throw out_of_range("length must be at least zero");
         }
 
-        mBuffer.resize((uint32_t)length);
+        mBuffer.resize((size_t)length);
     }
 
     int64_t MemoryStream::Position() const
@@ -102,25 +102,25 @@ namespace Lupus {
             throw out_of_range("position exceeds stream boundaries");
         }
 
-        advance((mIterator = begin(mBuffer)), (int32_t)position);
+        advance((mIterator = begin(mBuffer)), (int)position);
     }
 
-    int32_t MemoryStream::Read(vector<uint8_t>& buffer, uint32_t offset, uint32_t size)
+    int MemoryStream::Read(vector<uint8_t>& buffer, size_t offset, size_t size)
     {
         if (offset > buffer.size() || size > buffer.size() - offset) {
             throw out_of_range("offset and size does not match buffer size");
         } else if (mIterator >= end(mBuffer)) {
             return 0;
         } else if ((int64_t)size > (int64_t)mBuffer.size() - Position()) {
-            size = (uint32_t)((int64_t)mBuffer.size() - Position());
+            size = (size_t)((int64_t)mBuffer.size() - Position());
         }
 
         copy(mIterator, mIterator + size, begin(buffer) + offset + size);
         advance(mIterator, size);
-        return (int32_t)size;
+        return (int)size;
     }
 
-    int32_t MemoryStream::ReadByte()
+    int MemoryStream::ReadByte()
     {
         if (end(mBuffer) <= mIterator) {
             return -1;
@@ -129,19 +129,19 @@ namespace Lupus {
         return *(mIterator++);
     }
 
-    int32_t MemoryStream::Write(const vector<uint8_t>& buffer, uint32_t offset, uint32_t size)
+    int MemoryStream::Write(const vector<uint8_t>& buffer, size_t offset, size_t size)
     {
         if (!mWritable) {
             throw not_supported();
         } else if (offset > buffer.size() || size > buffer.size() - offset) {
             throw out_of_range("offset and size does not match buffer size");
-        } else if (size > (uint32_t)mBuffer.capacity()) {
+        } else if (size > (size_t)mBuffer.capacity()) {
             mBuffer.reserve(mBuffer.capacity() + (size > 1024 ? size : 1024));
         }
 
         mBuffer.insert(mBuffer.end(), begin(buffer) + offset, begin(buffer) + offset + size);
         mIterator = end(mBuffer);
-        return (int32_t)size;
+        return (int)size;
     }
 
     void MemoryStream::WriteByte(uint8_t byte)
@@ -157,27 +157,27 @@ namespace Lupus {
     {
         switch (origin) {
             case SeekOrigin::Begin:
-                advance((mIterator = begin(mBuffer)), (int32_t)offset);
+                advance((mIterator = begin(mBuffer)), (int)offset);
                 break;
 
             case SeekOrigin::Current:
-                mIterator += (int32_t)offset;
+                mIterator += (int)offset;
                 break;
 
             case SeekOrigin::End:
-                advance((mIterator = end(mBuffer)), (int32_t)offset);
+                advance((mIterator = end(mBuffer)), (int)offset);
                 break;
         }
 
         return Position();
     }
 
-    uint32_t MemoryStream::Capacity() const
+    size_t MemoryStream::Capacity() const
     {
-        return (uint32_t)mBuffer.capacity();
+        return (size_t)mBuffer.capacity();
     }
 
-    void MemoryStream::Capacity(uint32_t cap)
+    void MemoryStream::Capacity(size_t cap)
     {
         mBuffer.reserve(cap);
     }
@@ -191,12 +191,12 @@ namespace Lupus {
         return mBuffer;
     }
 
-    uint8_t& MemoryStream::operator[](uint32_t i)
+    uint8_t& MemoryStream::operator[](size_t i)
     {
         return mBuffer[i];
     }
     
-    const uint8_t& MemoryStream::operator[](uint32_t i) const
+    const uint8_t& MemoryStream::operator[](size_t i) const
     {
         return mBuffer[i];
     }
