+++ Source/LupusCore/MemoryStream.h
@@ -14,11 +14,11 @@ namespace Lupus {
 
         MemoryStream() = default;
         MemoryStream(const std::vector<uint8_t>&) NOEXCEPT;
-        MemoryStream(uint32_t size) NOEXCEPT;
+        MemoryStream(size_t size) NOEXCEPT;
         MemoryStream(const std::vector<uint8_t>&, bool writable) NOEXCEPT;
-        MemoryStream(const std::vector<uint8_t>&, uint32_t offset, uint32_t size) throw(std::out_of_range);
-        MemoryStream(const std::vector<uint8_t>&, uint32_t offset, uint32_t size, bool writable) throw(std::out_of_range);
-        MemoryStream(const std::vector<uint8_t>&, uint32_t offset, uint32_t size, bool writable, bool visible) throw(std::out_of_range);
+        MemoryStream(const std::vector<uint8_t>&, size_t offset, size_t size) throw(std::out_of_range);
+        MemoryStream(const std::vector<uint8_t>&, size_t offset, size_t size, bool writable) throw(std::out_of_range);
+        MemoryStream(const std::vector<uint8_t>&, size_t offset, size_t size, bool writable, bool visible) throw(std::out_of_range);
         virtual ~MemoryStream() = default;
 
         virtual bool CanRead() const NOEXCEPT override;
@@ -30,18 +30,18 @@ namespace Lupus {
         virtual void Length(int64_t) throw(std::out_of_range) override;
         virtual int64_t Position() const NOEXCEPT override;
         virtual void Position(int64_t) throw(std::out_of_range) override;
-        virtual int32_t Read(std::vector<uint8_t>& buffer, uint32_t offset, uint32_t size) throw(std::out_of_range) override;
-        virtual int32_t ReadByte() NOEXCEPT override;
-        virtual int32_t Write(const std::vector<uint8_t>& buffer, uint32_t offset, uint32_t size) throw(not_supported, std::out_of_range) override;
+        virtual int Read(std::vector<uint8_t>& buffer, size_t offset, size_t size) throw(std::out_of_range) override;
+        virtual int ReadByte() NOEXCEPT override;
+        virtual int Write(const std::vector<uint8_t>& buffer, size_t offset, size_t size) throw(not_supported, std::out_of_range) override;
         virtual void WriteByte(uint8_t byte) throw(not_supported) override;
         virtual int64_t Seek(int64_t offset, SeekOrigin origin) NOEXCEPT override;
 
-        virtual uint32_t Capacity() const;
-        virtual void Capacity(uint32_t);
+        virtual size_t Capacity() const;
+        virtual void Capacity(size_t);
         virtual const std::vector<uint8_t>& GetBuffer() const throw(unauthorized_access);
 
-        virtual uint8_t& operator[](uint32_t);
-        virtual const uint8_t& operator[](uint32_t) const;
+        virtual uint8_t& operator[](size_t);
+        virtual const uint8_t& operator[](size_t) const;
 
     private:
